requirements.txt
pillow==7.0.0
keras==2.4.3
tensorflow==2.2.0
flask==1.1.1
requests==2.24.0